# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: StockPilot - Inventory & POS SaaS
  description: '

    ## StockPilot API - Inventory & POS SaaS


    Multi-tenant inventory management, point-of-sale, finance tracking, and AI insights

    for small businesses.


    ### Authentication

    All API endpoints require a Bearer token in the Authorization header:

    `Authorization: Bearer <access_token>`


    Include the business ID in the X-Business-Id header:

    `X-Business-Id: <business_id>`


    ### Rate Limiting

    - Default: 100 requests per minute per user

    - Auth endpoints: 5 requests per minute per IP


    ### Support

    For issues and feature requests, contact support.

    '
  version: 1.0.0
paths:
...
  /api/v1/businesses/me:
    get:
      tags:
      - businesses
      summary: Get My Business
      operationId: get_my_business_api_v1_businesses_me_get
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - businesses
      summary: Update My Business
      operationId: update_my_business_api_v1_businesses_me_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/businesses:
    post:
      tags:
      - businesses
      summary: Create Business
      description: 'FR-2.1 / FR-24: authenticated owner can create an additional business profile.'
      operationId: create_business_api_v1_businesses_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/businesses/me/logo:
    post:
      tags:
      - businesses
      summary: Upload Logo
      description: 'FR-2.2: store business logo. Local filesystem initially.'
      operationId: upload_logo_api_v1_businesses_me_logo_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_logo_api_v1_businesses_me_logo_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees:
    get:
      tags:
      - employees
      summary: List Employees
      operationId: list_employees_api_v1_employees_get
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeOut'
                title: Response List Employees Api V1 Employees Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - employees
      summary: Create Employee
      operationId: create_employee_api_v1_employees_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{membership_id}/deactivate:
    post:
      tags:
      - employees
      summary: Deactivate
      operationId: deactivate_api_v1_employees__membership_id__deactivate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          title: Membership Id
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{membership_id}/activate:
    post:
      tags:
      - employees
      summary: Activate
      operationId: activate_api_v1_employees__membership_id__activate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          title: Membership Id
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{membership_id}/role:
    patch:
      tags:
      - employees
      summary: Update Role
      description: 'FR-3.3/FR-26.2: owner can re-assign employee roles.'
      operationId: update_role_api_v1_employees__membership_id__role_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          title: Membership Id
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRoleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{membership_id}/reset-password:
    post:
      tags:
      - employees
      summary: Reset Password
      description: 'FR-26.4: owner resets an employee''s password (then shares it out-of-band).'
      operationId: reset_password_api_v1_employees__membership_id__reset_password_post
      security:
      - HTTPBearer: []
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          title: Membership Id
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeResetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{membership_id}:
    delete:
      tags:
      - employees
      summary: Remove
      operationId: remove_api_v1_employees__membership_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          title: Membership Id
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_logo_api_v1_businesses_me_logo_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_logo_api_v1_businesses_me_logo_post
    BusinessCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
          default: BDT
        tax_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Rate
          default: 0.0
      type: object
      required:
      - name
      title: BusinessCreate
    BusinessOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        logo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Logo Path
        currency:
          type: string
          title: Currency
        tax_rate:
          type: number
          title: Tax Rate
        invoice_format:
          type: string
          title: Invoice Format
        min_stock_default:
          type: integer
          title: Min Stock Default
      type: object
      required:
      - id
      - name
      - address
      - phone
      - email
      - logo_path
      - currency
      - tax_rate
      - invoice_format
      - min_stock_default
      title: BusinessOut
    BusinessUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
        tax_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Rate
        invoice_format:
          anyOf:
          - type: string
          - type: 'null'
          title: Invoice Format
        min_stock_default:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Stock Default
      type: object
      title: BusinessUpdate
...
    EmployeeCreate:
      properties:
        name:
          type: string
          title: Name
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        role:
          type: string
          pattern: ^(Owner|Manager|Cashier)$
          title: Role
      type: object
      required:
      - name
      - password
      - role
      title: EmployeeCreate
    EmployeeOut:
      properties:
        membership_id:
          type: integer
          title: Membership Id
        user:
          $ref: '#/components/schemas/UserOut'
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - membership_id
      - user
      - role
      - is_active
      title: EmployeeOut
    EmployeeResetRequest:
      properties:
        new_password:
          type: string
          maxLength: 128
          minLength: 6
          title: New Password
      type: object
      required:
      - new_password
      title: EmployeeResetRequest
    EmployeeRoleUpdate:
      properties:
        role:
          type: string
          pattern: ^(Owner|Manager|Cashier)$
          title: Role
      type: object
      required:
      - role
      title: EmployeeRoleUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - id
      - name
      - email
      - phone
      - is_active
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer